backend/requirements.txt
fastapi
uvicorn
atlassian-python-api
pyyaml
apscheduler
python-dotenv
requests
pytz
pytest
httpx
